Drop get_checkin_without_comment_url from overlay.
Move tests to document overview and mail overview to test
VisibleActionButtonRendererMixin.
