+++ opengever/bumblebee/tests/test_overlay_adapter_mail.py
@@ -56,17 +56,6 @@ class TestGetFile(IntegrationTestCase):
         self.assertEqual(self.mail_eml.message, adapter.get_file())
 
 
-class TestGetCheckinWithoutCommentUrl(IntegrationTestCase):
-
-    features = ('bumblebee', )
-
-    def test_returns_none_because_its_not_possible_to_checkin_emails(self):
-        self.login(self.regular_user)
-        adapter = getMultiAdapter((self.mail_eml, self.request), IBumblebeeOverlay)
-
-        self.assertIsNone(adapter.get_checkin_without_comment_url())
-
-
 class TestGetCheckinWithCommentUrl(IntegrationTestCase):
 
     features = ('bumblebee', )
